Refactor: Extract Custom Type Casts related tests
